Change speed value to uint16_t

+++ src/MSP432LISAT/MSP432P401R_LISAT_ccs/SerialInterface/serialESPBridge.cpp
@@ -102,7 +102,7 @@ void processCommand() {
                 speedString += serialBuffer[i];
                 i++;
             }
-            int speedValue = stoi(speedString);
+            uint16_t speedValue = stoi(speedString);
             changeMotorSpeed(speedValue);
             break;
         }
